Remove deluxequeues references
As this project is based on an old version of DeluxeQueues, supports a different platform, and is tailored specifically to the RTGame server, it no longer makes sense for it to be branded as a DeluxeQueues fork.

I have therefore renamed it to ProxyQueues and detached the fork. The copyright notice remains for appropriate credit to the remaining DelxueQueues work.