Add queue pausing
Queues can now be paused to prevent attempts to connect queued players.

Queues can be paused using /queue pause or via the api in another plugin.

Multiple plugins can pause queues at once, with each providing a reason. A queue will only unpause when all pauses are removed.

Suitable use cases are situations where players are guaranteed to fail to connect to the server for a period of time, such as lockdowns or the server being offline.

Pausing a queue helps prevent situations where a player queueing for multiple servers is stuck trying to connect to a server that is not available, without attempting to join other queued servers.